chatbot-backend/requirements.txt
flask 
psycopg2-binary 
openai
python-dotenv
dateparser
numpy 
faiss-cpu
PyPDF2